requirements.txt
Package             Version
------------------- -------
asgiref             3.10.0
Django              4.2
djangorestframework 3.16.1
pillow              12.0.0
pip                 25.3
setuptools          65.5.0
sqlparse            0.5.3
typing_extensions   4.15.0
tzdata              2025.2
